1.3.2 XSS CSRF Fix (#12)
* updated readme

* xss fix/bumped versions

* added nonces to ajax responses

* added esc_attr

* added esc_html

+++ inc/wpeac-admin.php
@@ -82,6 +82,7 @@ class WPEAC_Admin {
 		<div class="wrap">
 			<h2>Cache Options</h2>
 			<form method="post" action="options.php">
+			<?php wp_nonce_field( 'wpeac-settings', 'wpeac-settings-nonce' ); ?>
 				<?php settings_fields( 'wpengine-advanced-cache-control' );?>
 				<?php do_settings_sections( 'wpengine-advanced-cache-control' ); ?>
 				<p>Increasing the cache times on the server will allow more users to see Cached copies of your pages. Cached copies of pages are served from outside of WordPress, which conserves server resources and saves time for your users.
@@ -148,7 +149,7 @@ class WPEAC_Admin {
 					</tr>
 				</table>
 				<?php submit_button(); ?>
-				<input type="hidden" id="wpe_ac_global_last_modified" name="<?php echo esc_attr( WPEAC_Core::CONFIG_OPTION . '[wpe_ac_global_last_modified]' ); ?>" value="<?php echo WPEAC_Core::get( 'wpe_ac_global_last_modified' )?>"/>
+				<input type="hidden" id="wpe_ac_global_last_modified" name="<?php echo esc_attr( WPEAC_Core::CONFIG_OPTION . '[wpe_ac_global_last_modified]' ); ?>" value="<?php echo esc_attr( WPEAC_Core::get( 'wpe_ac_global_last_modified' ) )?>"/>
 			</form>
 			<hr>
 			<table class="form-table">
@@ -266,6 +267,7 @@ class WPEAC_Admin {
 	 * @return null
 	 */
 	function purge_varnish_post_id_callback() {
+		check_admin_referer( 'wpeac-settings', 'wpeac-settings-nonce' );
 		$this->purge_cache_by_post_id( $_POST['your_post_id'] );
 		wp_die(); // this is required to terminate immediately and return a proper response
 	}
@@ -281,7 +283,7 @@ class WPEAC_Admin {
 	 */
 	function reset_global_last_modified_callback() {
 		$this->update_global_last_modified();
-		echo WPEAC_Core::get( 'wpe_ac_global_last_modified' );
+		echo esc_html( WPEAC_Core::get( 'wpe_ac_global_last_modified' ) );
 		wp_die(); // this is required to terminate immediately and return a proper response
 	}
 	/**
@@ -295,6 +297,7 @@ class WPEAC_Admin {
 	 * @return null
 	 */
 	function purge_varnish_path_callback() {
+		check_admin_referer( 'wpeac-settings', 'wpeac-settings-nonce' );
 		$this->purge_cache_by_path( $_POST['your_path'] );
 		wp_die(); // this is required to terminate immediately and return a proper response
 	}
@@ -473,7 +476,8 @@ class WPEAC_Admin {
 		} else {
 			$purge_response = "$post_id is not a valid public post ID";
 		}
-		echo $purge_response;
+		header( 'content-type: text/plain' );
+		echo esc_html( $purge_response );
 	}
 	/**
 	 * Returns post_types to manage
@@ -543,10 +547,12 @@ class WPEAC_Admin {
 			add_filter( 'wpe_purge_varnish_cache_paths', array( $this, 'set_path' ) );
 			WpeCommon::purge_varnish_cache( 1 );
 			remove_filter( 'wpe_purge_varnish_cache_paths', array( $this, 'set_path' ) );
-			echo ( "Purged $url from cache." );
+			$purge_path_response = "Purged $url from cache.";
 		} else {
-			echo 'This function only works on WP Engine installations.';
+			$purge_path_response = 'This function only works on WP Engine installations.';
 		}
+		header( 'content-type: text/plain' );
+		echo esc_html( $purge_path_response );
 	}
 
 	function set_path() {
